# gossip: Make boot process faster and safer