--- a/message/messaging_service.cc
+++ b/message/messaging_service.cc
@@ -449,6 +449,7 @@ static constexpr unsigned do_get_rpc_client_idx(messaging_verb verb) {
     case messaging_verb::GOSSIP_DIGEST_ACK2:
     case messaging_verb::GOSSIP_SHUTDOWN:
     case messaging_verb::GOSSIP_ECHO:
+    case messaging_verb::GOSSIP_QUERY_TOKEN_STATUS:
     case messaging_verb::GET_SCHEMA_VERSION:
         return 1;
     case messaging_verb::PREPARE_MESSAGE:
@@ -907,6 +908,17 @@ void messaging_service::unregister_gossip_echo() {
 }
 future<> messaging_service::send_gossip_echo(msg_addr id) {
     return send_message_timeout<void>(this, messaging_verb::GOSSIP_ECHO, std::move(id), 3000ms);
+}
+
+// GOSSIP_QUERY_TOKEN_STATUS
+void messaging_service::register_gossip_query_token_status(std::function<future<gms::gossip_query_token_status_response> (const rpc::client_info& cinfo)>&& func) {
+    register_handler(this, messaging_verb::GOSSIP_QUERY_TOKEN_STATUS, std::move(func));
+}
+void messaging_service::unregister_gossip_query_token_status() {
+    _rpc->unregister_handler(netw::messaging_verb::GOSSIP_QUERY_TOKEN_STATUS);
+}
+future<gms::gossip_query_token_status_response> messaging_service::send_gossip_query_token_status(msg_addr id) {
+    return send_message_timeout<gms::gossip_query_token_status_response>(this, messaging_verb::GOSSIP_QUERY_TOKEN_STATUS, std::move(id), 5000ms);
 }
 
 void messaging_service::register_gossip_shutdown(std::function<rpc::no_wait_type (inet_address from)>&& func) {