@asias 2019-07-24
Yes.

On Wed, Jul 24, 2019, 20:17 Benny Halevy <notifications@github.com> wrote:

> The dependency is on #4710 <https://github.com/scylladb/scylla/pull/4710>
> right?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/scylladb/scylla/pull/4756?email_source=notifications&email_token=AACOETGKYP7X7T7FT7DY2XLQBBB4HA5CNFSM4IGOMKX2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD2WEQUY#issuecomment-514607187>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AACOETCTBHIABYLWASRH5BDQBBB4HANCNFSM4IGOMKXQ>
> .
>