@tgrabiec 2019-07-26
> gossip: Make boot process faster and safer
> 
> During boot up, we need to wait for gossip in multiple places:
> 
> 1. Wait for gossip to settle to enable features. See commit
>    [71bf757](https://github.com/scylladb/scylla/commit/71bf757b2c8fb372bfd115ed00bf56069331d7cc) ("gossiper: Enable features
>    only after gossip is settled").
> 2. Wait until local node knows tokens of peer nodes before announcing
>    join status. To make sure when the writes sent by the other nodes for
>    pending token range arrives, the local node knows all the tokens of the
>    cluster. See commit [89b769a](https://github.com/scylladb/scylla/commit/89b769a073da1a1727b913159d85b97fe911676c).
> 3. Wait until peer nodes know the bootstrap tokens of local node before
>    streaming. To make sure all peer nodes knows the pending token range
>    before the streaming starts, so that all writes after streaming starts
>    that belong to the pending token range will be forwarded to the new node.
> 4. Wait until gossip is settled before start the cql server. See commit
>    [8c90912](https://github.com/scylladb/scylla/commit/8c909122a6a6b90ddb50ea8be250829c199616ab).
> 
> Before this patch, we sleep ring_delay which is 30 seconds by default
> and wait for the size of endpoint_state_map does not change over some
> time. The time out based approach does not guarantee the above
> conditions are met. It can wait foo much or wait too few.
> 
> This patch introduces a new method. Instead of waiting for some
> arbitrary time, we ask peer nodes to return the number of nodes it
> knows, the checksum of normal tokens, the checksum of bootstrap tokens
> and the checksum of gossip status (e.g., NORMAL), the checksum of nodes
> leaving the cluster. If the checksums between local node and all remote
> nodes are identical, it is guaranteed all nodes have the same view of
> tokens and nodes status.
> 
> A new RPC verb GOSSIP_QUERY_TOKEN_STATUS is introduced for this new
> method. If any of the nodes in the cluster does not support this new
> verb, or a node is down and can not respond the new verb, it will
> fallback to the old timeout based method. A nice thing is that we do not
> need to negotiate between nodes if GOSSIP_QUERY_TOKEN_STATUS is
> supported. When the node that supports GOSSIP_QUERY_TOKEN_STATUS send
> this verb to node that does not support it during boot process, the
> sender will get rpc::unknown_verb_error then it will fallback.

Why do we need to introduce a new RPC protocol for this, and reuse existing gossip protocol by forcing and waiting for gossip exchange with each of the peers (and peers we learned about)?

If you think it's ok to proceed immediately if the current node knows the same thing as other nodes (as this PR assumes), then you can force gossip rounds to make sure that this is always the case. Then there's no need to fall back to the slow time-based wait in case there is a disagreement since there will never be in the absence of communication failure.