@asias 2019-07-26
> Why do we need to introduce a new RPC protocol for this, rather than reuse existing gossip protocol by forcing and waiting for gossip exchange with each of the peers (and peers we learned about)?

This accelerates the gossip propagation. This patch does not plan to solve this problem. We still need a way to verify each node has the same view.

> If you think it's ok to proceed immediately if the current node knows the same thing as other nodes (as this PR assumes), then you can force gossip rounds to make sure that this is always the case. Then there's no need to fall back to the slow time-based wait in case there is a disagreement since there will never be in the absence of communication failure.

Making a node talking to all nodes does not guarantee this node know everything other nodes knows. E.g. n1 knows n1,n2,n3 but it does not know n4. When n1 talks (send gossip syn message) to n2 which knows n4, n2 will not send n4's info to n2. Only when n2 (which knows n4) talks to (send gossip syn message) n1, n1 will know n4. 

Also, what if a node failed to talk to a peer node in gossip? Don't we have the same issue like we can not send the GOSSIP_QUERY_TOKEN_STATUS verb to a remote node?